+++ examples/plot_background_model.py
@@ -6,20 +6,15 @@ from astropy.coordinates import Angle
 from gammapy.background import CubeBackgroundModel
 from gammapy import datasets
 
-# read
-filename = '../test_datasets/background/bg_cube_model_test.fits'
-filename = datasets.get_path(filename, location='remote')
+filename = datasets.get_path('../test_datasets/background/bg_cube_model_test.fits',
+                             location='remote')
 bg_model = CubeBackgroundModel.read(filename, format='table')
 
-# plot
 bg_model.plot_image(energy=Quantity(2., 'TeV'))
 bg_model.plot_spectrum(det=Angle([0., 0.], 'degree'))
 
-# write
 outname = 'cube_background_model'
-bg_model.write('{}_bin_table.fits'.format(outname), format='table',
-               write_kwargs=dict(clobber=True))
-bg_model.write('{}_image.fits'.format(outname), format='image',
-               write_kwargs=dict(clobber=True))
+bg_model.write('{}_bin_table.fits'.format(outname), format='table', clobber=True)
+bg_model.write('{}_image.fits'.format(outname), format='image', clobber=True)
 
 plt.show()