+++ CHANGES.rst
@@ -51,6 +51,8 @@ Pull requests
 - Function to plot Fermi 3FGL light curves [#286] (Jonathan Harris)
 - Add EffectiveAreaTable exporter to EffectiveAreaTable2D [#276] (Johannes King)
 - Add interface to HESS FitSpectrum JSON output [#296] (Christoph Deil)
+- Remove gammapy.shower package [#291] (Christoph Deil)
+-  Add cube background model class [#299] (Manuel Paz Arribas)
 
 .. _gammapy_0p2_release: